moved test to tests
