+++ hlib/python/hlib/frontend/tests/cifar-10.py
