+++ hlib/python/hlib/frontend/tests/test_numpy_func.py
