+++ hlib/python/hlib/frontend/tests/y_mnist.npy
